+++ ledger/participant-integration-api/src/test/suite/scala/platform/store/backend/DbDtoEqSpec.scala
@@ -1,0 +1,35 @@
+package com.daml.platform.store.backend
+
+import org.scalatest.matchers.should.Matchers
+import org.scalatest.wordspec.AnyWordSpec
+
+class DbDtoEqSpec extends AnyWordSpec with Matchers {
+
+  import DbDtoEq._
+
+  "DbDtoEq" should {
+
+    "compare DbDto when used with `decided` keyword" in {
+
+      val dto0 = DbDto.ConfigurationEntry(
+        ledger_offset = "",
+        recorded_at = 0,
+        submission_id = "",
+        typ = "",
+        configuration = Array[Byte](),
+        rejection_reason = None,
+      )
+
+      val dto1 = dto0.copy()
+      val dto2 = dto0.copy()
+
+      dto0 should equal(dto0) // Works due to object equality shortcut
+      dto1 shouldNot equal(dto2) // As equality is overridden to be false with DbDto
+      dto1 should equal(dto2)(decided by DbDtoEq)
+      List(dto1) should equal(List(dto2))(decided by DbDtoSeqEq)
+
+    }
+
+  }
+
+}
